+++ internals-js/src/values.ts
@@ -525,7 +525,7 @@ function isValidValueApplication(value: any, locationType: InputType, locationDe
     const valueKeys = new Set(Object.keys(value));
     const fieldsAreValid = locationType.fields().every(field => {
       valueKeys.delete(field.name);
-      return isValidValueApplication(value[field.name], field.type!, undefined, variableDefinitions)
+      return isValidValueApplication(value[field.name], field.type!, field.defaultValue, variableDefinitions)
     });
     const hasUnexpectedField = valueKeys.size !== 0
     return fieldsAreValid && !hasUnexpectedField;